simplify sync process
